requirements.txt
moviepy>=1.0.3
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=9.0.0
python-pptx>=0.6.20
scikit-image>=0.19.0

# 可選依賴
markitdown>=0.1.1  # 增強型Markdown生成功能
openai>=1.0.0      # AI輔助功能 